+++ crates/mimic/src/data/query/load/dynamic.rs
@@ -84,8 +84,8 @@ pub struct LoadQueryDyn {
     pub selector: Selector,
     pub format: LoadFormat,
     pub lookup: Option<Where>,
-    pub limit: Option<usize>,
-    pub offset: usize,
+    pub limit: Option<u32>,
+    pub offset: u32,
     pub include_children: bool,
 }
 
@@ -114,21 +114,21 @@ impl LoadQueryDyn {
 
     // offset
     #[must_use]
-    pub const fn offset(mut self, offset: usize) -> Self {
+    pub const fn offset(mut self, offset: u32) -> Self {
         self.offset = offset;
         self
     }
 
     // limit
     #[must_use]
-    pub const fn limit(mut self, limit: usize) -> Self {
+    pub const fn limit(mut self, limit: u32) -> Self {
         self.limit = Some(limit);
         self
     }
 
     // limit_option
     #[must_use]
-    pub const fn limit_option(mut self, limit: Option<usize>) -> Self {
+    pub const fn limit_option(mut self, limit: Option<u32>) -> Self {
         self.limit = limit;
         self
     }
